Create FoodController and AdminFoodcontroller

+++ backend-spring-boot/Online Food Ordering/Online-Food-Ordering/src/main/java/com/heli/request/CreateFoodRequest.java
@@ -16,7 +16,7 @@ public class CreateFoodRequest {
     private Category category;
     private List<String> images;
 
-    private Long restaurantId;
+    private int restaurantId;
     private boolean vegetarin;
     private boolean seasional;
     private List<IngredientsItem> ingredients;
